Use tags when linking to Github compare view
After updating the repo state from the compare endpoint, grab the sha of
the two ends of the compare, fetch the list of tags for the repo, look
up the first tag for each of the shas, and use those tags to build the
compare URL linked to from the title.

Our use case for this tracker is to compare the difference between a
branch that tracks what's been deployed to production and master.
Because these branches are moving targets, when we prepare a release
candidate for a deploy we link to specific commits which are tagged with
(never updated) release_123 type tags. Using those tags to build the
compare URL makes it one-click easy to get a link for a release
candidate.