Add Analytics
To get an idea of how many people are using this, and which of
features. Useful to know when making changes in the future.

Anonymise the  value as it's personal each user and could
be used to do nasty things to someones github account if they've
not set permissions properly.

The rest of the options are user specific, but nothing you couldn't
get from a public github page